Projekti3/requirement.txt
mysql.connector
pyjwt
pycrypto
cryptography